moving page class to its subdir
