+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:fluttertodo/home_page.dart';
+import 'package:fluttertodo/pages/home_page.dart';
 
 void main() => runApp(const MyApp());
